Diagnose uses of declarations that are too early.
There are lots of ways a declaration can be used before we have the
information necessary to handle that use. Issue diagnostics for these.

Interleave declaration and type-checking of global declarations so that
declaring a later declaration can depend on the results of type-checking
an earlier one.

Fixes #1394, fixes #1395, fixes #1396.