Don't duplicate rewrite constraints as equality constraints.
Instead, when checking that a constraint is satisfied, check both its
equality constraints and its rewrite constraints. This is a necessary
prerequisite for applying rewrite constraints to equality constraints,
where we would otherwise collapse the implied equality constraints to a
tautological `V == V` constraint.

In passing, make ErrorBuilder support building the error message more
incrementally and use that to improve diagnostics for mismatched values
with equality constraints.