Compute a fixed point of rewrite constraints.
When resolving a constraint type, compute the fixed point of the
specified rewrite constraints. If no such fixed point exists, detect
that situation and diagnose the problem.

The algorithm used here is to iteratively apply all rewrites to each
rewrite constraint until either one of them refers to itself, which
indicates there's a cycle in the rewrite graph, or the set of rewrites
converges.

This algorithm has some pathological inputs in which the runtime can
grow exponentially in the size of the input (indeed, the fully-rewritten
set of rewrites can grow exponentially in the size of the input, so this
is unavoidable), so an iteration limit is also provided.