Rephrased all "you" instances into a communal "we"
A higher sense of community, belonging and accountability is achieved when using "we" to talk about us as a community, including when addressing issues. So I suggest to use that phrasing throughout the CoC. 
Then, the "we" used in the previous version for the conduct team needed to also be explicited into "conduct team" or "team". So I updated these instances as well.