Make type contexts expect a value of type Type.
In particular, this means that a type can implement `ImplicitAs(Type)`
and have values of that type behave like types.

This implies that `()` and `{}` are no longer types. They are now values
whose type is the result of converting `()` or `{}` to type `Type`. This
fixes a lot of cases where these types were previously mishandled.