explorer: number support for 8bit and 16bit integers
This is probably implemented very poorly, it was done over two days in a attempt to work myself into the codebase and understand it.
Its done in a way where the core principle of the int type is not modified.
What happens is that the type int can have a size and when a value is stored its reduced to the wanted byte size,
when used within the parser its expanded to a 4byte signed integer again, basically wasting that space, but since the various apis expect
a int32_t it was the easiest way to integrate this without a major refactor of the types used.

This way might seam somewhat hacky and id argue it is but it solves the problem.