Add `-march=native` to our compile flags.
This makes the native CPU instructions available. I noticed that we
weren't already doing this when working on another bit of code where its
actually relevant. This doesn't make a big difference for any of the few
benchmarks we have at the moment, but it seems like a good idea.

Modern Clang/LLVM support this exact spelling both on x86 and ARM CPUs,
so its surprisingly portable. I've tested it on my ARM mac just in case.