README.md: add a section with past talks (#2434)
Suggesting to add a section with past conference talks, to include content that has been created around carbon. This feel to me like an accessible type of content to have in the main README for those wanting to dig a bit deeper without going into /docs.

This does not show the speaker name to keep it "Carbon/community" oriented, vs naming, but there's an argument both ways I think.