Hack `clang-format` config to format our macros better.
This somewhat abuses control-flow macro handling in `clang-format` to
get the behavior we want. Fortunately, I don't think we're ever likely
to want this for its intended purpose so it seems harmless to co-opt it
like this. If we really wanted, we can narrow it to `if`-macros, but
I picked the simpler option to start.

FWIW, I tried all the other macro formatting special cases to see if one
would work but it didn't. Might be worth filing a feature request to get
a `TypeDefinitionMacros` setting to compliment `TypenameMacros` and
format like this does, but it seems (very) low priority.